test(ui): split story testbed cli coverage
